Make Matcher.all multi-platform
