+++ kotest-assertions/kotest-assertions-core/src/jvmMain/kotlin/io/kotest/matchers/reflection/compose.kt
@@ -4,7 +4,7 @@ import io.kotest.matchers.Matcher
 import io.kotest.matchers.MatcherResult
 import kotlin.reflect.KProperty1
 
-@Deprecated("Replaced with io.kotest.matchers.compose.Matcher.all Deprecated since 5.6")
+@Deprecated("Replaced with io.kotest.matchers.compose.Matcher.all Deprecated since 5.7")
 @Suppress("UNCHECKED_CAST")
 fun <T : Any?> Matcher.Companion.compose(
    vararg pairs: Pair<Matcher<*>, KProperty1<T, *>>
